Add files via upload
This is a client created in order to easily navigate the ( https://restcountries.eu/ ) API. It supports all the features that this API supports, including filtered queries. All the results appear in new tabs so that the user won't have to go back and forth, and they also appear in an organized way.